chore: layer card refactor
