--- src/components/layers/basemaps/styles/BasemapCard.module.css
@@ -1,19 +1,0 @@
-.card {
-    position: relative;
-    margin: 8px 4px 8px 4px;
-    padding-bottom: 0;
-}
-
-.visibility {
-    height: 32px;
-    width: 32px;
-    padding: 4px;
-}
-
-.content {
-    padding: 0;
-}
-
-.tooltip {
-    margin-top: -8px;
-}
