--- src/components/layers/layers/styles/LayerCard.module.css
@@ -1,60 +1,0 @@
-.card {
-    position: relative;
-    margin: 8px 4px 0 4px;
-    padding-bottom: 0;
-    z-index: 2000;
-}
-
-.header {
-    height: 54px;
-    padding: 2px 8px 0 18px;
-    font-size: 14px;
-}
-
-.title {
-    width: 227px;
-    padding-left: 15px;
-    font-size: 15px;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    font-weight: 500px;
-    line-height: 17px;
-}
-
-.subheader {
-    width: 195px;
-    padding-left: 15px;
-    line-height: 17px;
-    font-size: 14px;
-}
-
-.legend {
-    padding-left: 32px;
-}
-
-.actions {
-    background-color: #eee;
-    height: 32px;
-}
-
-.visibility {
-    position: absolute;
-    right: 28px;
-    top: 4px;
-}
-
-.expand {
-    position: absolute;
-    right: -4px;
-    top: 4px;
-}
-
-.content {
-    font-size: 14px;
-    padding: 0;
-}
-
-.tooltip {
-    margin-top: -8px;
-}
