+++ src/components/layers/overlays/AddLayerButton.js
