+++ src/components/layers/overlays/styles/OverlayCard.module.css
@@ -1,0 +1,3 @@
+.legend {
+    padding-left: 32px;
+}
