+++ src/components/layers/styles/LayerCard.module.css
@@ -1,3 +1,9 @@
+.card {
+    position: relative;
+    margin: 8px 4px 0 4px;
+    padding-bottom: 0;
+}
+
 .cardHeader {
     height: 54px;
     padding: 2px 8px 0 18px;
@@ -8,11 +14,11 @@
     color: var(--colors-grey900);
 }
 
-.content {
+.title {
     flex: 1 1 auto;
 }
 
-.content h2 {
+.title h2 {
     width: 242px;
     font-size: 15px;
     white-space: nowrap;
@@ -23,12 +29,17 @@
     margin: 0;
 }
 
-.content h3 {
+.title h3 {
     line-height: 17px;
     font-size: 14px;
     font-weight: 400;
     margin: 0;
     color: var(--colors-grey800);
+}
+
+.sortable h2,
+.sortable h3 {
+    padding-left: 15px;
 }
 
 .action {
@@ -46,4 +57,8 @@
     height: 48px;
     padding: 12px;
     cursor: pointer;
+}
+
+.content {
+    font-size: 14px;
 }
