+++ src/components/layers/styles/SortableHandle.module.css
@@ -1,0 +1,9 @@
+.handle {
+    width: 24px;
+    height: 24px;
+    position: absolute;
+    left: 6px;
+    top: 16px;
+    cursor: ns-resize;
+    fill: #333;
+}
